Rrubric scoring model; change points type to double, update validation logic, and enhance handling of max scores in rubric criteria

+++ Server/src/main/java/com/grade/forge/grading/service/SubmissionGradeService.java
@@ -217,7 +217,7 @@ public class SubmissionGradeService {
         if (score < 0) {
             throw new IllegalArgumentException("awardedScore cannot be negative");
         }
-        Integer maxAllowed = subCriteria.getMaxScore();
+        Double maxAllowed = subCriteria.getMaxScore();
         if (maxAllowed != null && score > maxAllowed) {
             throw new IllegalArgumentException("awardedScore cannot exceed sub-criteria maxScore");
         }
